add authorization

+++ controllers/deleteUser.js
@@ -5,6 +5,11 @@ const pool = require('../dbConnection');
 const router = express.Router();
 
 router.delete('/:id', (req, res) => {
+    if (!(req.role === '1')) {
+        res.status(403).send('Forbidden');
+        return;
+    }
+
     pool.query(`delete from employee where id = ${req.params.id}`, (error) => {
         if (error) throw error;
